feat(breadbox): Error reporting from celery worker (#275)
