+++ breadbox/breadbox/celery_task/utils.py
@@ -111,28 +111,23 @@ def format_task_status(task):
             # sometimes we get an error while the celery state is still progress
             task._state = TaskState.FAILURE.name
             if isinstance(task._result, UserError):
-                message = str(task.result.detail)
+                message = str(task._result.detail)
         if isinstance(task.result, UserError):
             # this is a specific, expected error that we check for
             # return error message for the front to display
             message = str(task.result.detail)
         elif isinstance(task.result, HTTPException):
-            raise task.result
-            # message = {
-            #     "status_code": str(task.result.status_code),
-            #     "detail": str(task.result.status_code),
-            # }
+            message = {
+                "status_code": str(task.result.status_code),
+                "detail": str(task.result.status_code),
+            }
         elif isinstance(task.result, FileValidationError):
-            raise task.result
-            # message = str(task.result)
+            message = str(task.result)
         else:
-            # this is an unexpected error, somewhere in our code threw that exception
-            # task.result contains the exception. throw it to throw a hard 500 and report to stackdriver
-            print("-----")
-            print(type(task.result))
-            print(task.result)
-            print("-----")
-            raise CeleryException("Error from celery worker.") from task.result
+            # This is an unexpected error thrown while the task was running. 
+            # At this point, the error has already been logged in the celery error reporter 
+            # and should be visible in the GCS Error Groups.
+            message = "Encountered an unexpected error. Please try again later."
     elif task.state == TaskState.PENDING.name:
         # pending means we have not entered the task yet
         pass
