+++ breadbox/breadbox/main.py
@@ -1,14 +1,12 @@
-from .startup import create_app, ensure_directories_exist, GCPExceptionReporter
+from .startup import create_app, ensure_directories_exist
 from .celery_task.utils import create_celery
-from .config import get_settings, Settings
+from .config import get_settings
 from breadbox.api.dependencies import get_user
 from fastapi import Request, status
-from fastapi.exceptions import RequestValidationError
 from fastapi.exception_handlers import (
-    request_validation_exception_handler,
     http_exception_handler,
 )
-from .logging_config import configure_logging
+from .logging import configure_logging, GCPExceptionReporter
 
 configure_logging()
 
@@ -26,7 +24,7 @@ celery_app = create_celery()
 celery = celery_app
 
 # create the Google Cloud exception reporter class
-exception_reporter = GCPExceptionReporter(settings.breadbox_env)
+exception_reporter = GCPExceptionReporter(service="breadbox", env_name=settings.breadbox_env)
 
 from starlette.exceptions import HTTPException as StarletteHTTPException
